Normalize JSON whitespace in Asgard example
This patch applies a practice used for the CASE JSON-LD examples
currently undergoing migration from CASE 0.1, normalizing JSON-LD
whitespace according to the output of this command:

    python3 -m json.tool in.json out.json

Only whitespace changes are introduced in this patch.

This patch partially addresses CASE Jira Task ONT-248.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions to this
toolkit is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.