Add /vendor/bundle directory to excludes
With all due irony, adding the `excludes` directive to support the
contribute file broke the script the contribute file provided.

The local website runner in `contribute.md` creates a directory
`vendor/bundle` at the top of the repository.  This directory was
ignored until the `exclude` was added to `_config.yml`, apparently
causing the default excludes list (commented out for illustration in
`_config.yml`) to be overwritten instead of augmented.   Jekyll started
to fail serving the local page because of an automatically generated
blog page in that bundle directory.

This patch removes the vendor/bundle directory from the website build
scan.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>