Remove incorrectly-placed fileSystemType properties
There is value in recording the type of file system in which a file was
found.  This is not a property on a File facet.  While Sean and I agreed
to not want to present an incorrect representation, we disagreed on how
this should be represented, and found it best to remove the property for
now.

Acked-by: Sean Barnum <sbarnum@mitre.org>