Add Makefile to synchronize TSV files
To back the generated table contents, TSV files need to reside under
`/_data/`, per this page:
https://jekyllrb.com/docs/datafiles/
(I'm by no means a Jekyll expert, so it's possible there's another
mechanism that can be used to refer to files in the same directory.
Pointers are welcome.)

To help confirm files remain synchronized, a `make check` rule is also
included here, and has been set up to be called from a top-level call to
`make check` after a `make clean`.  This maintenance step is now
documented in `/CONTRIBUTING.md`.

I have confirmed this commit passes the freshly added `make check`.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of these maintenance instructions
is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>