Use Makefile to derive CSV files
I received a request to have the machine-readable files be available as
CSV and TSV files.  This Makefile addition creates a build chain to
ensure the CSV files are synchronized against the TSV files that back
the generated webpage content.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of these maintenance instructions
is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>