Provide standalone SPARQL query runner for testing data
This patch lays groundwork for implementing the JSON-LD for the
"Urgent Evidence" narrative, being tracked in Jira as Story ONT-107.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of these maintenance instructions
is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software or toolkit is necessarily the best
available for the purpose.

References:
* [ONT-107] - ONT-44-JSON-LD

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>