Add Multiple Browsers narrative to Example Gallery
With the current gallery design, the URLs and Browsing History topic
needs description text to link this narrative.  For the sake of posting
narratives to the website, the narrative is being posted now, and will
be linked when the topic text is written.

This commit ports text drafted by CASE community members.

This patch resolves ONT-218.

References:
* [ONT-218] ONT-189-Narrative-MultipleBrowsers
* [ONT-324] Add "URLs and Browsing History" topic text to gallery page
  to link narratives

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>