Add Insider Threat narrative to Example Gallery
With the current gallery design, the File System topic needs description
text to link this narrative.  For the sake of posting narratives to the
website, the narrative is being posted now, and will be linked when the
topic text is written.

This commit ports text drafted by CASE community members.

This patch resolves ONT-236.

References:
* [ONT-236] ONT-185-Narrative-InsiderThreat
* [ONT-328] Add "File System" topic text to gallery page to link
  narratives

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>