Add Password-Protected Files narrative to Example Gallery
With the current gallery design, the Passwords, Tokens, Credentials
topic needs description text to link this narrative.  For the sake of
posting narratives to the website, the narrative is being posted now,
and will be linked when the topic text is written.

This commit ports text drafted by CASE community members.

This patch resolves ONT-237.

References:
* [ONT-237] ONT-188-Narrative-PasswordProtectedFiles
* [ONT-330] Add "Passwords, Tokens, Credentials" topic text to gallery
  page to link narratives

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>