Add WhatsApp Messages narrative to Example Gallery
With the current gallery design, the Messages topic needs description
text to link this narrative.  For the sake of posting narratives to the
website, the narrative is being posted now, and will be linked when the
topic text is written.

This commit ports text drafted by CASE community members.

This patch resolves ONT-273.

References:
* [ONT-273] ONT-150-Narrative-WhatsAppMessages
* [ONT-331] Add "Messages" topic text to gallery page to link narratives

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>