Add Distribution narrative to Example Gallery
With the current gallery design, the Installed Apps topic needs
description text to link this narrative.  For the sake of posting
narratives to the website, the narrative is being posted now, and will
be linked when the topic text is written.

This commit ports text drafted by CASE community members.

This patch resolves ONT-302.

References:
* [ONT-302] ONT-187-Narrative-Distribution
* [ONT-333] Add "Installed Apps" topic text to gallery page to link
  narratives

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>