Build JSON-LD of "Asgard" example into separate file
This patch takes the JSON-LD that was hard-coded in the HTML of the
"Asgard" example page, and breaks it out into separate JSON files that
compile into both the HTML page and a self-contained JSON-LD file.

The Makefile recursion from the top directory was updated to descend
into examples directory, and then into each example with a Makefile.

This accomplishes a part of ONT-129, assisting with reviewing for the
CASE 0.2.0 release (ONT-352).

This patch includes a confirmation that the updated
`make clean && make && make check` path work as written for the
examples/ directory.  I unfortunately found that I'd forgotten the
middle `make` in the just-merged commit for ONT-353, though, and will be
updating that separately.

References:
* [ONT-129] JSON-LD examples on the website may need to be
  machine-readable
* [ONT-352] Test website examples against release-0.2.0 branch
* [ONT-356] Prepare "Asgard" JSON-LD for querying

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>