Update Asgard example for CASE release-0.2.0
This patch squash-commits the changes made to the Asgard example in the
ONT-352 branch, except for linking to the asgard.json file.  I would
prefer to provide that file only when it can be easily synchronized with
the JSON rendered in the HTML.

ONT-352 is not being merged yet because it builds on ONT-356, and 356
will take some discussion before merging.

References:
* [ONT-231] (CP-10) Definition of Trace from CASE prototype not
  preserved
* [ONT-352] Test website examples against release-0.2.0 branch
* [ONT-356] Prepare "Asgard" JSON-LD for querying

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>