Add Github Action for Continuous Integration checks
This workflow executes steps I have been using to confirm data-driven
files, built for serving on Github's Jekyll service, are synchronized
with their source files.  Actual instructions to run tests in each
directory were encoded in Makefiles prior to this patch.

This patch addresses part of ONT-360: Establishing the CI framework.
A follow-on patch will add narrative page building, somewhat abstracted,
to the Make build chain.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions for this
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [ONT-360] The website repository needs Continuous Integration for MVP
  development

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>