Add virtual environment and Makefile descent into examples
This patch prepares the examples directory to be added to the Continuous
Integration build system for the website.

This patch guarantees that when `make` is executed from the top
directory, Make will descend into the examples directory after building
a virtual environment.  The virtual environment practice is to provide
rdflib for running SPARQL queries without requiring sudo to affect the
Python environment managed by the root user.  (The "requests" library is
included as an invisible dependency of rdflib, seeming to be a
distribution bug - a component of rdflib fails on import if requests is
not present.)

This patch does not add any build directives for anything in the
examples directory, but both ONT-107 and ONT-356 have plans to add Make
steps to build JSON-driven website pages and keep those pages
synchronized against source files.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions for this
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [ONT-107] MVP JSON-LD for Chain of Custody (Topic ONT-44)
* [ONT-356] Prepare "Asgard" JSON-LD for querying
* [ONT-360] The website repository needs Continuous Integration for MVP
  development

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>