Add query runner to assist with narrative page construction
This patch adds a script that will be used, at least initially, to run
SPARQL queries against JSON-LD data for the purpose of generating web
page components.  ONT-107 and ONT-309 provide the first instance of
using this, but for review, I wanted to separate out adding the packages
for the Continuous Integration.

Please note this disclaimer related to mention of specific software:
Participation by NIST in the creation of the instructions for this
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [ONT-107] MVP JSON-LD for Chain of Custody (Topic ONT-44)
* [ONT-309] Query - action timeline (Chain of Custody, Topic ONT-44)
* [ONT-360] The website repository needs Continuous Integration for MVP
  development

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>