Adjust kiosk-processing action timing
Eoghan turned the kiosk processing action into a time interval.  This
patch adjusts the time so the sort order for the timeline query is
maintained.  (When not hand-writing the data, timestamp granularity will
handle this just fine.)

This patch also adjusts the timeline table structure to include an
optional start time, also flattening the SPARQL Turtle syntax a bit so
the facet identifier can be re-used.  Removing the anonymous node
component should also make the SPARQL more accessible to those who don't
work with Turtle often.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>