Build Owl Trafficking page as Asgard page
The narratives on the website have code portions in separate files so
they can be consumed by test and query processes.  This patch starts
moving the JSON into separate files for testing, as was done for the
Asgard example.  The Asgard Makefile is ported in this patch, though for
the sake of passing CI and git-bisect the `check` recipe doesn't test
JSON syntax yet.  The Makefile build chain for the /examples directory
has the Owl Trafficking page integrated.

One technology upgrade since the Asgard example is generating a `sed`
script instead of laboriously copy-pasting `sed` commands.

This patch is confirmed to regenerate the index.html page for the Owl
Trafficking narrative, without changes.

References:
* [ONT-129] JSON-LD examples on the website may need to be
  machine-readable
* [ONT-349] Add "Owl" scenario to website, either in gallery or
  alongside Asgard
* [ONT-356] Prepare "Asgard" JSON-LD for querying

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>