Construct Owl Trafficking gestalt JSON-LD file
This updates some of the JSON components' style to make aggregating the
piecemeal JSON files possible without relying on an RDF engine.  The
style decisions are:
* Adding a `@graph` wrapper JSON key to the baseline `_0` file.
* Making each JSON file after `_0` have an array as its top-level data
  structure.

The Makefile now runs a script (ported from the Asgard directory) to
construct the gestalt JSON-LD file and test that the construction is
consistent with the components loaded into the web page.

The constructed JSON-LD file is now linked as a downloadable resource.

References:
* [ONT-129] JSON-LD examples on the website may need to be
  machine-readable
* [ONT-349] Add "Owl" scenario to website, either in gallery or
  alongside Asgard

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>