Adjust render of xsd:hexBinary data
On correcting the namespace prefix usage for AC-154, several XSD
literal-types began to be interpreted with stronger typing than strings.
The timestamps now render with numeric time zones.  The hexBinary data
was rendering as binary data, in a manner not desired for displaying
hashes.

The resolution to the hexBinary display issue was originally used in a
development branch (that I wrote) of the CASE-to-DFXML bidirectional
mapping repository.

References:
* [AC-154] XML Schema datatype prefix needs to be explicit in all
  examples
* [CASE-Implementation-DFXML] https://github.com/ajnelson-nist/case-implementation-dfxml

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>