Correct kindOfRelationship literal references
This patch aligns literals with those defined in UCO's vocabulary
namespace.

Literals declared to be in the namespace, but that are not in the
namespace, have had their types changed to xsd:string.

These literals were identified with the OC-56 branch's updates from
today on the CASE-Examples-QC repository.  These updates reduce the
undefined relationship literals to only instances of xsd:string that are
not in UCO's vocabulary.

References:
* [AC-164] Assess current usage of kindOfRelationship literals
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC/
* [UCO OC-56] (CP-49) UCO needs to document the core:Relationship object
  and why it is used instead of RDF predicates

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>