Track CASE-Utilities-Python as submodule
This patch series sets up importing the sample SPARQL running script as
the command case_sparql_select, instead of relying on the script being
a website-housed feature.

This patch series also sets up importing the command
case_sparql_construct, necessary to implement the query in ONT-387.

Patch 1 ensures the submodule is cloned as part of the build proces, and
that new submodule management artifacts are not exposed to Git tracking
or to website serving.

References:
* [AC-182] Have website track CASE-Utilities-Python as submodule
* [ONT-387] Query - actions taken to derive artifact (Urgent Evidence,
  Narrative ONT-118)

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>