Use case_sparql_select in query build chain
_config.yml no longer excludes the /src directory, because this commit
removes the /src directory.

setuptools is now also upgraded in the virtual environment as a result
of a practice found necessary at times when doing pip installs from
local file trees.

sample_sparql_runner commands are now replaced with case_sparql_select
commands.  Note that the argument order is reversed from how the
original runner script was written, to allow for feeding an arbitrary
number of graph files on the command line.  This use case will become
apparent with the implementation of ONT-387.

References:
* [AC-182] Have website track CASE-Utilities-Python as submodule
* [ONT-387] Query - actions taken to derive artifact (Urgent Evidence,
  Narrative ONT-118)

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>