Addresses ONT-431, 'rootExhibitNumber'
This commit illustrates 'case-investigation:rootExhibitNumber' by using the
urgent_evidence example, selecting initial objects in a ProvenanceRecord
chain.

Index.html.in updated to display the question, query, and results.
Makefile updated to automatically populate the index.html.in file.
The sparql file attached includes the query and is required for make.