Bump CASE-Utilities-Python to 0.3.0
CASE-Utilities-Python (`case_utils`) 0.3.0 obviates the need for the
website to track a direct submodule pointer to the CASE ontology, as
`case_utils` has its own submodule pointer to the CASE ontology, and
also provides a pre-built monolithic ontology file to use for
validation.

The `case_utils` submodule is still recursively initialized, to provide
access to `rdf-toolkit.jar`.

References:
* [AC-207] Release CASE-Utilities-Python 0.3.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>