Specify some ObservableObject subclasses
These subclass additions were done in support of AC-9.  The refinements
are mostly focused on the storage system hierarchy.  Other
`ObservableObject` non-subclassed instances remain available for
refinement.

This patch handles the manual revisions necessary for unit tests to
continue to work.  A follow-on patch will regenerate the Make-managed
files.

References:
* [AC-9] Publish NIST DFXML Implementation

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>