Confirm Turtle translation ability
The CASE-Examples-QC repository includes a step that converts example
JSON-LD to Turtle.  This had not yet been run when testing the upgrade
of rdf-toolkit to the post-1.10 state.  Not having done so, we missed
that there is a potential build-halting issue with `rdf-toolkit`.

The current version of `rdf-toolkit` contains a behavior (possibly a
bug) that dislikes this snippet of JSON-LD:

```json
{
    "@id": "urn:example:kb:object-1",
    "@type": "owl:Thing",
    "uco-core:hasFacet": {
        "@type": "uco-core:Facet"
    }
}
```

The blank node that contains only a type assertion raises a
`NullPointerException` in a step related to determining sort order of
blank nodes.

A follow-on patch will remove one occurrence of this behavior, as it
turns out to be unhelpful data (an empty `DeviceFacet`).  Before the
`ObservableObject` subclass hierarchy was populated, that empty `Facet`
was needed for duck-typing.  With `ObservableObject` subclasses in
place, it is no longer necessary.

This problem will need to be addressed upstream.

This patch revises the testing workflow to include a step converting the
glommed JSON-LD to Turtle for `pytest` review.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>