Fix Urgent Evidence queries and cut TODO
A few things happen in this patch, which all were related enough to
cause `pytest` to fail:

* The dangling TODO on cutting the 'unstable' pointer should have been
cut a while ago, but was forgotten.  This had an unintended interaction
with rdflib's 6.2.0 release.  With that release, namespace prefixes are
now decided by *last* `graph.parse` call.  So, the CASE Investigation
namespace prefix was being changed from `case-investigation:` (as in the
data graphs) to `investigation:` (as in the Unstable graph).
   - Cutting the unstable-build load meant the subclass hierarchy was
     needed to let the exhibit photos query work (due to its usage of
     the `a/rdfs:subClassOf*` query pattern in lieu of RDFS inferencing.
* Review of the `prepareQuery` calls was showing that these query
objects do not need to be constructed.  If something were being
parameterized, like a SPARQL variable binding, it would still be worth
using `prepareQuery`.  But these usages can get by with just passing a
string into `graph.query`.
* Removing the `prepareQuery` call obviated `nsdict`.

After the simplifications, `pytest` now passes in the `urgent_evidence`
directory.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>