Correct class reference
`uco-identity:Role` does not exist.  `uco-role:Role` does, but CASE has
a more specific subclass that would be more appropriate to use here.

This commit is only a correction of a class reference.  Further
discussion needs to be had on whether `role:Role` is appropriate to use
as an object of the `action:performer` property.  I believe this usage
is incongruous with other ontologies outside of UCO.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>