Remove incomplete hashes
A uco-types:Hash is required to specify its algorithm and value.  Some
of the example data is missing a value, and/or leaving a placeholder in
place.  Since this will only raise a validation error until a complete
Hash object is provided, this patch cuts the placeholders for the sake
of reducing validation issues to 0.

Note that rdf-toolkit has a bug in processing a blank node that only has
a type designation.  So, in one case where a `ContentDataFacet` was left
empty, the `ContentDataFacet` was cut to avoid breaking CI.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>