Add identifiers for hash nodes
The UUIDv5 form is drawn from the 2003 draft Thiemann hash URN
namespace.  The following Python call generated the UUID v5 for
Asgard's `kb:hash-bfa02feb-6184-5ca5-9ba3-272bf4367c5d`:

```python
uuid.uuid5(
    uuid.NAMESPACE_URL,
    "urn:hash::sha256:7ea081166336119da78ee4bbdbd06840b94efe28988a2bdb0bcf2387a481e283".lower()
)
```

Note this UUID generation pattern is an option, but not necessarily a
best practice for the community.  Users must be aware of the tradeoffs
between node identifier correlation versus risks from hash preimage
analysis, as well as UUID v5's inherent reliance on SHA-1.

For CASE's hand-drafted examples, the maintainers find the value of
finding copy-paste errors in drafting outweighs the known risks.

The prior patch saw that some `Hash` objects had incongruous pairings of
algorithm and value-length.  As UUIDv5s were generated for this patch,
the algorithm in the URN was assumed to be the one corresponding with
the value length, rather than the one recorded in the example.

A follow-on patch will regenerate Make-managed files.

References:
* https://datatracker.ietf.org/doc/html/draft-thiemann-hash-urn-01
  (Note this document expired 2004-03-04.)

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>