De-duplicate IRIs
When SHACL-reviewing all of the example graphs of the CASE website
concatenated together, some IRIs were found to have been copied and
pasted from example to example with some properties changed.  In
particular, `uco-core:name`, required to be at most one value for an
IRI, pointed out the IRI repurposing when it was found to have two
values in the aggregate.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>