Revise Make descent order
This applies a Make pattern of depending on a phony Make target instead
of a flag file, to guarantee dependents of the monolothic ontology
builds are visited after the builds.  This has been used in CASE-Corpora
for some time, and also helps avoid an issue where updating the Git
submodules (such as in bumping the prerelease pointers) necessitated
manually deleting the flag file.

No effects were observed on Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>