Disambiguate UUID used in URL, URLVisit, and mapped tool data
The UUID `39ff4987-8ae5-47e3-8369-dbd0d5f79398` appeared in the mapped
tool data.  The CASE graph forwarded it to a URL node, and then it
appears there were subsequent copies of the UUID as other example data
was drafted.

This patch removes all occurrences of the UUID from the CASE graph(s),
because it is not clear (to me) whether the mapped object from the
tool's data model corresponds to a UCO `URL` object.  It more appears
(to me) to correspond to a `URLHistory` object.

This was among shared UUID usages found while reviewing CASE-Examples-QC
PR 54.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ajnelson-nist/CASE-Examples-QC/pull/54

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>